-Changed weight to a double to support fractional weights

+++ app/src/main/java/com/example/nathan/myapplication/Move.java
@@ -2,7 +2,7 @@ package com.example.nathan.myapplication;
 
 public abstract class Move {
     Position endPoint;
-    int weight;
+    double weight;
     String description;
 
 //    public Move() {
